+++ src/discljord/specs.clj
@@ -46,6 +46,16 @@
 (s/def ::query string?)
 (s/def ::limit number?)
 
+(s/def ::name string?)
+(s/def ::type (s/or :keyword #{:game :stream :music}
+                    :int int?))
+(s/def ::activity (s/keys :req-un [::name ::type]
+                          :opt-un [::url]))
+
+(s/def ::idle-since number?)
+(s/def ::status #{:online :offline :invisible :idle :dnd})
+(s/def ::afk boolean?)
+
 ;; -------------------------------------------------
 ;; discljord.messaging specs